+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.4] - 2020-04-19
+### Added
+- `ByteChannelCoAcceptor`: sends data to a Kotlin `ByteWriteChannel`
+- `CoStrings.kt`: various functions to handle character output
+
 ## [0.3] - 2020-02-18
 ### Added
 - `ChannelCoAcceptor`: sends data to a Kotlin `Channel`